+++ packages/site-demo/content/product/components/user-block/react/size-m.tsx
@@ -1,19 +1,14 @@
 import { Size, UserBlock } from '@lumx/react';
 import React from 'react';
 
-export const App = ({ theme }: any) => {
-    const logAction = (action: string) => () => console.log(action);
-
-    return (
-        <UserBlock
-            theme={theme}
-            name="Emmitt O. Lum"
-            fields={['Creative developer', 'Denpasar']}
-            avatarProps={{ image: '/demo-assets/persona.png', alt: 'Avatar' }}
-            size={Size.m}
-            onMouseEnter={logAction('Mouse entered')}
-            onMouseLeave={logAction('Mouse left')}
-            onClick={logAction('UserBlock clicked')}
-        />
-    );
-};
+export const App = ({ theme }: any) => (
+    <UserBlock
+        theme={theme}
+        name="Emmitt O. Lum"
+        nameProps={{ 'aria-label': 'Emmitt O. Lum - open user profile' }}
+        fields={['Creative developer', 'Denpasar']}
+        avatarProps={{ image: '/demo-assets/persona.png' }}
+        size={Size.m}
+        onClick={console.log}
+    />
+);